Added Marius version of better Conv3D
Conv_3D_big is the version that has 99% training acc on PCA, but performs better than the previous best model in any case.

Conv_3D_big_gated is the same version now with an additional GLU-Gate added after Convulution. This performed better on 32x32 Images, but seemed to reduce accs for PCA. Handle with care.

Note: None of these have been run yet, so they may include error and will be updated if that is the case.